SDC-3445: Multiple pipelines writing to MapR FS errors
There are two important pieces of knowledge to understand what is going
on:

* Since MapR have native libraries that can't be loaded in one JVM
  twice, all MapR components (MaprFS included) are running with shared
  classloader (privateClassLoader = false). On the contrary all
  other hadoop stages (HDFS, CDH, Apache) are running with private
  classloader set to true.

* HDFS call FileSystem.get() that we're using to retrieve FileSystem
  object (that is then used for all HDFS operations) is using a
  classloader wide cache.

Outcome of those two together means that we were essentially reusing the
same FileSystem object in all MapR FS components in all pipelines that
are running in a single SDC instance (single JVM).

This is a problem because we're (as the nice citizens) cleaning up unused
resources including the shared FileSystem object by calling a close() on
it. So while starting or stopping pipeline, we could close FileSystem
object that is still used by a another running pipeline. Which explains
why we've seen "stream closed" exceptions.

Rather then using FileSystem.get(), I've change the HDFS Target to
use FileSystem.newInstance() that is guaranteed to return a new
FileSystem object (will always bypass the cache). I've also refactored
the HDFS Target to keep this created instance for the entire execution
of pipeline and reuse it for all FS operations.

Change-Id: I785350ab9112c9e538dee8cf76b253c1a04e56e5
Reviewed-on: https://review.streamsets.net/3715
Reviewed-by: Junko Urata <junko@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>